# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Arceus Internal MCP API
  version: "1.0.0"
  summary: Fastify-backed internal control plane for agent heartbeats.
  description: |
    Headless API consumed by Arceus agents. All routes live under
    `/api/internal/v1/*`, require bearer auth, and accept an
    `Idempotency-Key` header for safe replay.
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://arceus.internal
    description: Internal staging
security:
  - BearerAuth: []
...
paths:
  /api/internal/v1/tasks:
    post:
      tags: [tasks]
      summary: Create a task
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/BeatId"
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/AgentRole"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTaskBody" }
      responses:
        "201":
          description: Task created
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskCreatedEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Validation" }
  /api/internal/v1/tasks/{taskId}:
    patch:
      tags: [tasks]
      summary: Partially update a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchTaskBody" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Validation" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
    BeatId:
      name: x-beat-id
      in: header
      required: true
      schema: { type: string }
    CompanyId:
      name: x-company-id
      in: header
      required: true
      schema: { type: string }
    AgentRole:
      name: x-agent-role
      in: header
      required: true
      schema:
        type: string
        enum: [ceo, cto, pm, developer, tester, ui_designer, marketing, skills_lead]
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Ok:
      description: Successful response
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessEnvelope" }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: Idempotency conflict or invalid state
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Validation:
      description: Request body failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationErrorEnvelope" }
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, summary]
      properties:
        status: { type: string, enum: [success] }
        summary: { type: string }
        data: { type: object, additionalProperties: true }
        nextActions:
          type: array
          items: { type: string }
...
    ErrorEnvelope:
      type: object
      required: [status, summary, error]
      properties:
        status: { type: string, enum: [error] }
        summary: { type: string }
        error:
          type: object
          required: [cause, retry, stopWhen]
          properties:
            cause:
              type: string
              enum: [validation, governance, not_found, conflict, upstream, internal]
            retry:
              type: string
              enum: [safe, unsafe, never]
            stopWhen: { type: string }
    ValidationErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/ErrorEnvelope"
        - type: object
          properties:
            error:
              type: object
              properties:
                details:
                  type: array
                  items:
                    type: object
                    properties:
                      field: { type: string }
                      message: { type: string }
                      code: { type: string }
    CreateTaskBody:
      type: object
      required: [title]
      properties:
        id: { type: string }
        title: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        problemStatement: { type: string, maxLength: 4000 }
        deliverable: { type: string, maxLength: 2000 }
        definitionOfDone:
          type: array
          items: { type: string }
        kind: { type: string }
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignedRole: { type: string }
        sprintId: { type: string, nullable: true }
        parentTaskId: { type: string, nullable: true }
        dependsOnTaskIds:
          type: array
          items: { type: string }
    PatchTaskBody:
      type: object
      minProperties: 1
      properties:
        title: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignedRole: { type: string }
        assignedAgentId: { type: string, nullable: true }
    TaskCreatedEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [taskId, status]
              properties:
                taskId: { type: string }
                status: { type: string }
...